requirements.txt
azureml-sdk==1.1.5
jsonschema==3.2.0
